+++ examples/react/src/index.js
@@ -18,6 +18,7 @@ const App = () => {
         show={show}
         src={src}
         onClose={() => { toggle(false) }}
+        onOpen={() => console.log('Editor is opened.')}
         onComplete={(props) => { console.log(props) }}
         onBeforeComplete={(props) => { console.log(props); return false; }}
       />