+++ projects/react/utils/index.js
@@ -9,3 +9,4 @@ export * from './get-icon';
 export * from './get-label';
 export * from './get-secret-header-name';
 export * from './get-watermark-position';
+export * from './get-image-sealing-params';