+++ CHANGELOG.md
@@ -25,6 +25,25 @@ Types of changes:
 - ...
 
 -------------
+## 3.9.1 - 2020-06-12
+### Fixed
+- onComplete method doesn't call with new callback syntax
+
+
+## 3.9.0 - 2020-06-12
+
+### Fixed
+- calculating final size when zooming was applied on download
+- problem with query sting on generate cloudimage url
+- configuration of tools
+- fix canvas initilization error
+
+### Added
+- image sealing
+- support for onOpen & onClose methods
+- fullscreen mode
+- focus point
+
 
 ## 3.7.7 - 2020-05-17
 ## 3.7.6 - 2020-05-17